public-eye-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-cloud-vision==3.5.0
supabase==2.0.3
httpx==0.24.1
Pillow==10.1.0
gunicorn==21.2.0
requests==2.31.0